+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added platform constant VK play.
 - Added player statistics.
 - Screen mode: portrait/landscapes.
+- Notification of other windows.
 
 # [v1.1.2](https://github.com/megalanthus/defold-gamepush/releases/tag/v1.1.2) (2023-04-05)